+++ react/typings/global.d.ts
@@ -199,6 +199,7 @@ interface RenderComponent<P={}, S={}> {
     ExtensionPoint: typeof ExtensionPoint
     Link: typeof Link
     NoSSR: any
+    Loading: any
     Helmet: any
     canUseDOM: boolean
     withHMR: any