+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Auth treatment and redirect to the login page.
+- Loading component to Render Exports
 
 ## [7.11.3] - 2018-6-21
 ### Removed
-- Remove support for root extension without component
+- Remove support for root extension without component.
 
 ## [7.11.2] - 2018-6-21
 ### Fixed